+++ cpskin/core/browser/controlpanel.py
@@ -139,6 +139,17 @@ class CPSkinControlPanelAdapter(SchemaAdapterBase):
         setShowSlogan,
     )
 
+    def getshowPortletsInRightActionsPanel(self):
+        return self.settings.show_portlets_in_right_actions_panel
+
+    def setshowPortletsInRightActionsPanel(self, value):
+        self.settings.show_portlets_in_right_actions_panel = value
+
+    show_portlets_in_right_actions_panel = property(
+        getshowPortletsInRightActionsPanel,
+        setshowPortletsInRightActionsPanel,
+    )
+
 
 class CPSkinControlPanel(ControlPanelForm):